refactor: move Viewmodels to Models
