+++ Lab5/Lab5/Models/LabViewModel.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Lab5.Models;
+
+public class LabViewModel
+{
+    [Required(ErrorMessage = "InputData is required.")]
+    public string InputData { get; set; } = string.Empty;
+}
