+++ Lab5/Lab5/Models/UserProfileViewModel.cs
@@ -1,4 +1,4 @@
-namespace Lab5.ViewModels;
+namespace Lab5.Models;
 
 public class UserProfileViewModel
 {
