--- Lab5/Lab5/ViewModels/LabViewModel.cs
@@ -1,6 +1,0 @@
-namespace Lab5.ViewModels;
-
-public class LabViewModel
-{
-    public string InputData { get; set; } = string.Empty;
-}
