+++ Lab5/Lab5/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-@model Lab5.ViewModels.UserLoginViewModel
+@model Lab5.Models.UserLoginViewModel
 
 @{
     ViewData["Title"] = "Login";
