+++ Lab5/Lab5/Views/Account/Profile.cshtml
@@ -1,4 +1,4 @@
-﻿@model Lab5.ViewModels.UserProfileViewModel
+﻿@model Lab5.Models.UserProfileViewModel
 @{
     ViewData["Title"] = "User Profile";
 }
