+++ Lab5/Lab5/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-@model Lab5.ViewModels.UserRegisterViewModel
+@model Lab5.Models.UserRegisterViewModel
 
 @{
     ViewData["Title"] = "Registration";
