+++ Lab5/Lab5/Views/Lab/Lab1.cshtml
@@ -1,4 +1,4 @@
-@model Lab5.ViewModels.LabViewModel;
+@model Lab5.Models.LabViewModel;
 @{
     ViewData["Title"] = "Lab1 Page";
 }
